pyproject.toml
[tool.poetry]
name = "atom-client"
version = "1.0.0"
description = "A client library for accessing atom"
authors = []
readme = "README.md"
packages = [
    {include = "atom_client"},
]
include = ["CHANGELOG.md", "atom_client/py.typed"]


[tool.poetry.dependencies]
python = "^3.9"
httpx = ">=0.20.0,<0.29.0"
attrs = ">=22.2.0"
python-dateutil = "^2.8.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["F", "I", "UP"]
